Adding checks to cleanup

+++ vantage6-server/vantage6/server/controller/cleanup.py
@@ -3,6 +3,7 @@ import logging
 from datetime import datetime, timedelta, timezone
 
 from vantage6.common.task_status import TaskStatus
+from vantage6.common.globals import DataStorageUsed
 from vantage6.server.model import Run
 from vantage6.server.model.base import DatabaseSessionManager
 from vantage6.server.service.azure_storage_service import AzureStorageService
@@ -46,7 +47,9 @@ def cleanup_runs_data(days: int, azure_config, include_input: bool = False):
             )
 
             for run in runs:
-                storage_adapter.delete_blob(run.result)
+                if run.result is not None and run.data_storage_used == DataStorageUsed.AZURE:
+                    log.debug(f"Deleting blob: {run.result}")
+                    storage_adapter.delete_blob(run.result)
                 run.result = ""
                 if include_input:
                     run.input = ""
